hook up return routing for websocket connections
