+++ agent/indy_catalyst_agent/transport/inbound/base.py
@@ -2,6 +2,8 @@
 
 from abc import ABC, abstractmethod
 
+from ...error import BaseError
+
 
 class BaseInboundTransport(ABC):
     """Base inbound transport class."""
@@ -9,3 +11,7 @@ class BaseInboundTransport(ABC):
     @abstractmethod
     def start(self) -> None:
         """Start listening for on this transport."""
+
+
+class TransportSetupError(BaseError):
+    """Setup error for an inbound transport."""
