+++ agent/indy_catalyst_agent/transport/inbound/manager.py
@@ -18,7 +18,7 @@ class InboundTransportManager:
 
         self.transports = []
 
-    def register(self, module_path, host, port, message_handler):
+    def register(self, module_path, host, port, message_handler, register_socket):
         """
         Register transport module.
 
@@ -27,6 +27,7 @@ class InboundTransportManager:
             host: The host to register on
             port: The port to register on
             message_handler: The message handler for incoming messages
+            register_socket: A coroutine for registering a new socket
 
         """
         try:
@@ -35,7 +36,9 @@ class InboundTransportManager:
             self.logger.warning(f"Failed to load module {module_path}")
             return
 
-        self.transports.append(imported_class(host, port, message_handler))
+        self.transports.append(
+            imported_class(host, port, message_handler, register_socket)
+        )
 
     async def start_all(self):
         """Start all registered transports."""
